Increase embedding threshold for fastembed 0.2.7 compatibility
In `test_examples_included_in_prompts` and `test_examples_included_in_pr≈ompts_2`
we make an HTP call to GET /api/models/qdrant/all-MiniLM-L6-v2-onnx to download
an embedding model to look up the most similar user intent to the user saying "Hi".

Our pyproject.toml has:

```
fastembed = [
  { version = ">=0.2.2, <0.5", python = ">=3.9,<3.10" },
  { version = ">=0.5.0, <=0.6.0", python = ">=3.10, <3.14" },
]
```

In Python 3.9.21 poetry instaled fastembed 0.2.7, giving a similarity of [(0.9502324275672436, 'hi')]
In Python 3.11.11 we use fastembed 0.6.0, giving a similairty of [(0.7792390137910843, 'hi')]

Prior to this commit, the threshold was set at 0.8. This means for Python 3.9.21
(fastembed 0.2.7) no LLM call was made (0.95>0.8). For fastembed 0.6.0 the same embedding
model gives 0.78 < 0.8. There's an issue here since the same embedding model should give
the same results regardless of the API/toolkit we use to call it.

But a bigger issue is that we should be mocking the fastembed call anyway
so we can return a value above, below, and exactly at the threshold and test each.
We shouldn't need internet access for unit-tests. Will file an issue for this.